Moved implementation into core
