+++ logbook-api/src/main/java/org/zalando/logbook/BodyReplacer.java
@@ -11,10 +11,6 @@ public interface BodyReplacer<T extends BaseHttpMessage> {
     @Nullable
     String replace(final T message);
 
-    static <T extends BaseHttpMessage> BodyReplacer<T> replaceBody(final Predicate<T> predicate, final String replacement) {
-        return message -> predicate.test(message) ? replacement : null;
-    }
-
     @SafeVarargs
     static <T extends BaseHttpMessage> BodyReplacer<T> compound(final BodyReplacer<T>... replacers) {
         return message -> Arrays.stream(replacers)
